+++ bobtemplates/odoo/hooks.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 # Copyright © 2016 ACSONE SA/NV
 # Copyright 2019 Anvar Kildebekov <https://it-projects.info/team/fedoranvar>
 # Copyright 2022 Yves Goldberg Ygol InternetWork
@@ -8,9 +7,15 @@ import ast
 import os
 import re
 
+import requests
 from mrbob.bobexceptions import ValidationError
-from mrbob.hooks import show_message
+from mrbob.hooks import show_message, to_boolean
 from pkg_resources import parse_version
+
+OCA_GITHUB_URL = (
+    "https://api.github.com/repos/OCA/maintainer-tools"
+    "/contents/template/module/readme"
+)
 
 
 def _dotted_to_camelcased(dotted):
@@ -74,7 +79,7 @@ def _add_local_import(configurator, package, module):
     init_path = os.path.join(configurator.target_directory, package, "__init__.py")
     import_string = "from . import {}".format(module)
     if os.path.exists(init_path):
-        with open(init_path, "U") as f:
+        with open(init_path) as f:
             init = f.read()
     else:
         init = ""
@@ -86,6 +91,21 @@ def _rm_suffix(suffix, configurator, path):
     path = os.path.join(configurator.target_directory, path)
     assert path.endswith(suffix)
     os.rename(path, path[: -len(suffix)])
+
+
+def _get_oca_readme_fragments(configurator, addon_name):
+    data = requests.get(OCA_GITHUB_URL).json()
+    for node in data:
+        if node["type"] != "file":
+            continue
+        file_name = node["name"]
+        file_url = node["download_url"]
+        r = requests.get(file_url)
+        path = os.path.join(
+            configurator.target_directory, addon_name + "readme", file_name
+        )
+        with open(path, "wb") as f:
+            f.write(r.content)
 
 
 #
@@ -149,22 +169,58 @@ def post_render_model(configurator):
 
 
 #
-# addon hooks
+# readme hooks
 #
 
 
-def pre_render_addon(configurator):
+def pre_render_readme(configurator):
     variables = configurator.variables
-    variables["odoo.version"] = int(variables["addon.version"].split(".")[0])
     variables["addon.name_camelwords"] = _underscored_to_camelwords(
         variables["addon.name"]
     )
 
 
+def post_render_readme(configurator, addon_name=""):
+    variables = configurator.variables
+    addon_name = addon_name + "/" if addon_name else addon_name
+    oca = (
+        variables["addon.oca"]
+        if isinstance(variables["addon.oca"], bool)
+        else to_boolean(False, False, variables["addon.oca"])
+    )
+    generate_readme = variables.get("addon.readme", True)
+    generate_readme = (
+        generate_readme
+        if isinstance(generate_readme, bool)
+        else to_boolean(False, False, generate_readme)
+    )
+    if oca and generate_readme:
+        _get_oca_readme_fragments(configurator, addon_name)
+    if oca or not generate_readme:
+        _delete_file(configurator, addon_name + "README.rst")
+    _delete_file(configurator, addon_name + "readme/dummy.txt")
+
+
+#
+# addon hooks
+#
+
+
+def pre_render_addon(configurator):
+    pre_render_readme(configurator)
+    variables = configurator.variables
+    variables["odoo.version"] = int(variables["addon.version"].split(".")[0])
+
+
 def post_render_addon(configurator):
     variables = configurator.variables
-    if variables["addon.oca"]:
-        _rm_suffix(".oca", configurator, variables["addon.name"] + "/README.rst.oca")
+    post_render_readme(configurator, addon_name=variables["addon.name"])
+    oca = (
+        variables["addon.oca"]
+        if isinstance(variables["addon.oca"], bool)
+        else to_boolean(False, False, variables["addon.oca"])
+    )
+    if oca:
         _rm_suffix(
             ".oca",
             configurator,
@@ -182,7 +238,6 @@ def post_render_addon(configurator):
             configurator, variables["addon.name"] + "/static/description/icon.png.oca"
         )
     else:
-        _delete_file(configurator, variables["addon.name"] + "/README.rst.oca")
         _delete_file(
             configurator, variables["addon.name"] + "/static/description/icon.png.oca"
         )